dead-letter-queue/requirements.txt
attrs==23.2.0
certifi==2022.9.24
cffi==1.16.0
charset-normalizer==2.1.1
dnspython==2.6.1
idna==3.4
kafka-python==2.0.2
python-dotenv==1.0.1
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
urllib3==1.26.13